initial commit
A working program that will record time-trace data and write to files as specified. You can specify a threshold so that when the standard devation of the data does not cross that threshold, then the data is considered as boring and only a single average value is recorded, otherwise a full detailed trace section will be recorded.